Add iterator-style API
The current API of argpar is parse-all-in-one-shot style: you pass
argc, argv and the argument descriptors, and it returns you with an
array of everything it was able to parse, or an error.

It would sometimes be useful for the client code to have a bit more
control on when to stop the argument parsing.  For example, let's
imagine a command line of this style:

  some-tool.exe some-condition --condition-opt1 --condition-opt2 \
      some-action-1 --action-1-opt \
      some-action-2 --action-2-opt

We would like to parse the option arguments to the `some-condition`
sub-command.  A strategy to do this using the current API would be to
call argpar_parse with descriptors for `--condition-opt1` and
`--condition-opt2`, and fail_on_unknown_opt=false.  argpar would stop
parsing after `some-action-1`.  We could then consume the items in the
returned array and stop when we see the non-option argument
`some-action-1`.

But let's consider the following error case:

  some-tool.exe some-condition --condition-opt1 --condition-opt2 --invalid-opt \
      some-action-1 --action-1-opt \
      some-action-2 --action-2-opt

argpar would stop parsing just after `--condition-opt2`.  The client
then needs to check the first non-ingested argument, see if it looks
like an option (starts with a `-`).  If it does (like in this case, with
`--invalid-opt`), it must emit its own "unrecognized option" message.

This works, but I don't find it very nice because the application needs
to do some work that the argument parser is expected to do.

For sophisticated use cases like this one, it would be useful for the
application to have more fine-grained control in when to stop the
parsing.  For this reason, this patch introduces a lower-level
iterator-style API.

To use it, the application first creates an argpar_state:

    struct argpar_state *state = argpar_state_create(argc, argv, descrs);

It then consumes a single argument at the time by repeatedly calling
argpar_state_parse_next:

    struct argpar_item *item;
    char *error;
    enum argpar_state_parse_next_status status;

    status = argpar_state_parse_next(state, &item, &error);
    while (insert_favorite_condition_here) {
	do_something_with_argpar_item();

        status = argpar_state_parse_next(state, &item, &error);
    }

In the success case above, the application would stop parsing upon
encountering the non-option argument `some-action-1`.

In the failure case above, the application would stop parsing when
argpar_state_parse_next returns
ARGPAR_STATE_PARSE_NEXT_STATUS_ERROR_UNKNOWN_OPT, when encountering
`--invalid-opt`.  It will have an error message ready to print in
`*error`.

Signed-off-by: Simon Marchi <simon.marchi@efficios.com>